requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
pymssql==2.2.5
python-dotenv==0.19.0
openpyxl==3.0.10
gunicorn==20.1.0